+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[Unreleased]]
 ### Added
+- Support for the [XLS-20 NFT proposal](https://github.com/XRPLF/XRPL-Standards/discussions/46)
+- `xrpl.models.amounts.get_amount_value` helper function
 - `xrpl.utils.str_to_hex` and `xrpl.utils.hex_to_str` helpers
 - `ledger_index` optional param for all the main account methods
 - `TicketCreate` transaction model